requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==5.0.0
billiard==4.2.1
boto3==1.40.5
botocore==1.40.5
cachetools==5.5.2
celery==5.4.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.3
distro==1.9.0
Django==4.2.7
django-appconf==1.1.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-googledrive-storage==1.5.2
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx==0.2.4
exceptiongroup==1.3.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.12.0
jmespath==1.0.1
jsonfield==3.1.0
kombu==5.5.4
lxml==6.0.2
msal==1.28.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.8.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
paramiko==3.0.0
pillow==11.3.0
prompt_toolkit==3.0.51
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
redis==5.1.1
reportlab==4.4.5
requests==2.32.4
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9.1
s3transfer==0.13.1
schedule==1.2.2
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
wsproto==1.2.0